+++ app/presenters/journey_presenter.rb
@@ -1,4 +1,4 @@
-class JourneyPresenter < SimpleDelegator
+class JourneyPresenter < BasePresenter
   # @return [Array<SectionPresenter>]
   def sections
     @sections ||= sections_with_tasks.map do |section|
@@ -13,11 +13,6 @@ class JourneyPresenter < SimpleDelegator
 
   def description
     category.title
-  end
-
-  # @return [String]
-  def created_at
-    super.strftime("%e %B %Y")
   end
 
   # Return the specification in HTML