+++ spec/factories/support/case.rb
@@ -5,9 +5,16 @@ FactoryBot.define do
     state { 0 }
     support_level { 0 }
 
-    association :enquiry, factory: :support_enquiry
     association :category, factory: :support_category
     sub_category_string { "category subtitle" }
+
+    trait :open do
+      state { :open }
+    end
+
+    trait :resolved do
+      state { :resolved }
+    end
 
     trait :with_documents do
       transient do
@@ -15,7 +22,7 @@ FactoryBot.define do
       end
 
       after(:create) do |kase, evaluator|
-        create_list(:support_document, evaluator.document_count, documentable: kase)
+        create_list(:support_document, evaluator.document_count, case: kase)
         kase.reload
       end
     end