+++ spec/features/support/case_summary_spec.rb
@@ -75,8 +75,8 @@ RSpec.feature "Case summary" do
   context "when the case is created" do
     it "has action links" do
       within "ul.govuk-list" do
-        expect(page).to have_link "Assign to case worker", href: "/support/cases/#{support_case.id}/edit", class: "govuk-link"
-        expect(page).to have_link "Resolve case", href: "#", class: "govuk-link"
+        expect(page).to have_link "Assign to case worker", href: "/support/cases/#{support_case.id}/assignment/new", class: "govuk-link"
+        expect(page).to have_link "Resolve case", href: "/support/cases/#{support_case.id}/resolution/new", class: "govuk-link"
       end
     end
   end
@@ -86,9 +86,9 @@ RSpec.feature "Case summary" do
 
     it "has action links" do
       within "ul.govuk-list" do
-        expect(page).to have_link "Change case owner", href: "/support/cases/#{support_case.id}/edit", class: "govuk-link"
+        expect(page).to have_link "Change case owner", href: "/support/cases/#{support_case.id}/assignment/new", class: "govuk-link"
         expect(page).to have_link "Add a case note", href: "/support/cases/#{support_case.id}/interactions/new?option=note", class: "govuk-link"
-        expect(page).to have_link "Send email", href: "#", class: "govuk-link"
+        expect(page).to have_link "Send email", href: "/support/cases/#{support_case.id}/email/type/new", class: "govuk-link"
         expect(page).to have_link "Log contact with school", href: "/support/cases/#{support_case.id}/interactions/new?option=contact", class: "govuk-link"
       end
     end